requirements.txt
# Core
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2

# APIs
betfairlightweight==2.18.0

# Database
psycopg2-binary==2.9.9

# Scheduler
APScheduler==3.10.4

# ML/Stats
scikit-learn==1.3.2
scipy==1.11.4

# Interface
streamlit==1.29.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

redis==5.0.1

rich==13.9.4

openai==1.54.3

fastapi==0.115.6

uvicorn==0.34.0
pydantic==2.10.5